requirements.txt
falcon==3.1.1
jsonschema
python-rapidjson
uvicorn
asyncpg
python-dotenv